+++ pkg/controller/queuejobresources/service/service.go
@@ -131,7 +131,7 @@ func (qjrService *QueueJobResService) deleteService(obj interface{}) {
 }
 
 
-func (qjrService *QueueJobResService) GetAggregatedResourcesByPriority(priority int, job *arbv1.AppWrapper) *clusterstateapi.Resource {
+func (qjrService *QueueJobResService) GetAggregatedResourcesByPriority(priority float64, job *arbv1.AppWrapper) *clusterstateapi.Resource {
         total := clusterstateapi.EmptyResource()
         return total
 }
